+++ crate/object_model/src/config/sprite/sprite_sheet_definition.rs
@@ -11,17 +11,17 @@ pub struct SpriteSheetDefinition {
     /// we store a `PathBuf`, it would need to re-allocate another `String`.
     pub path: String,
     /// Width of each individual sprite on the sprite sheet.
-    pub sprite_w: f32,
+    pub sprite_w: u32,
     /// Height of each individual sprite on the sprite sheet.
-    pub sprite_h: f32,
+    pub sprite_h: u32,
     /// Number of rows in the sprite sheet.
     ///
     /// This is the number of sprites counting down the sheet.
-    pub row_count: usize,
+    pub row_count: u32,
     /// Number of columns in the sprite sheet.
     ///
     /// This is the number of sprites counting across the sheet.
-    pub column_count: usize,
+    pub column_count: u32,
     /// Whether or not there is a 1 pixel border between sprites.
     #[serde(default = "SpriteSheetDefinition::default_has_border")]
     pub has_border: bool,